修改输出数据类型为long，可转换数最大升至九千九百九十九亿九千九百九十九万九千九百九十九

+++ src/CHS2Num/Chinese2Number.java
@@ -11,7 +11,7 @@ public class Chinese2Number {
     /**输入字符串**/
     private String inText;
     /**输出结果**/
-    private int out;
+    private long out;
     private boolean minus = false;
 
     /***
@@ -174,8 +174,8 @@ public class Chinese2Number {
      * @param part
      * @return
      */
-    private int parthandle(String part){
-        int result = 0;
+    private long parthandle(String part){
+        long result = 0;
         if(part.startsWith("十")){
             part="一"+part;
         }
